# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SIGET — API de Gestión de Maquinaria y Operaciones
  description: |
    API REST para la gestión de operaciones diarias de la flota vehicular, autenticación de usuarios
    y parámetros generales del sistema.

    ## Autenticación

    La mayoría de los endpoints requieren un token **JWT**. Flujo recomendado:

    1. Llamar a `POST /auth/login` con `email` y `password` para obtener un `access_token`.
    2. Enviar dicho token en la cabecera `Authorization: Bearer {access_token}` en cada petición protegida.
    3. Cuando el token esté por expirar, llamar a `POST /auth/refresh` (con el token actual todavía vigente)
       para obtener uno nuevo.
    4. Llamar a `POST /auth/logout` para invalidar el token de forma explícita.

    ## Formato de errores

    - **422 Unprocessable Content**: error de validación. Responde `{ "success": false, "message": "...", "errors": { "campo": ["mensaje"] } }`.
    - **401 Unauthorized**: token ausente, inválido o expirado.
    - **403 Forbidden**: el usuario está autenticado pero no tiene permiso sobre el recurso solicitado.
    - **404 Not Found**: el recurso indicado en la ruta no existe.
    - **500 Internal Server Error**: error inesperado del servidor. Responde `{ "message": "...", "error": "..." }`.

    ## Usuarios de prueba

    Estas cuentas están precargadas en el entorno de pruebas y sirven para probar la
    API con cada rol desde `POST /auth/login`:

...
  version: "1.9.0"
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /vales:
    get:
      tags: [Vales]
      summary: Listar vales de combustible
      description: |
        Lista paginada de vales de combustible, ordenados por `nro_vale` y `gestion`
        descendente.

        **Siempre** queda restringida a vales pendientes y vigentes (a diferencia de
        `GET /vales/pendientes`, que expone el mismo comportamiento en un formato distinto, este
        endpoint sí acepta filtros adicionales). Esto agrega siempre, sin importar los filtros
        recibidos, las condiciones `estado_vale = PENDIENTE` **y** `fecha_vencimiento >= ahora`.

        Las restricciones de visibilidad dependen del rol del usuario autenticado:
        - **conductor** (sin ningún otro rol): solo ve sus propios vales (`id_conductor` =
          persona del usuario).
        - **jefe-area** (puro, o combinado con `conductor`, pero sin `administrador`/
          `super-admin`): ve todos los vales de vehículos asignados a alguna de sus áreas a
          cargo, sin importar quién los emitió o a nombre de qué conductor.
        - **administrador** / **super-admin** (incluso combinado con otros roles): sin
          restricción, ve todos los vales del sistema.
      parameters:
        - name: nro_vale
          in: query
          description: |
            Filtra por número de vale. Compara tanto contra `nro_vale` de forma exacta
            como contra la combinación `nro_vale/gestion` (p.ej. `"2/2026"`).
          schema: { type: string }
          example: "2"
        - name: fecha_desde
          in: query
          description: Filtra vales cuya `fecha_emision` sea igual o posterior a esta fecha.
          schema: { type: string, format: date }
          example: "2026-08-01"
        - name: fecha_hasta
          in: query
          description: Filtra vales cuya `fecha_emision` sea igual o anterior a esta fecha.
          schema: { type: string, format: date }
          example: "2026-08-31"
        - name: estado_vale
          in: query
          description: |
            Filtra por estado exacto del vale. **En la práctica solo `PENDIENTE` puede devolver
            resultados**: como el endpoint ya restringe siempre a `estado_vale = PENDIENTE` (ver
            arriba), enviar `USADO` o `ANULADO` aquí combina ambas condiciones con `AND` y la
            consulta nunca encuentra coincidencias (lista vacía).
          schema:
            type: string
            enum: [PENDIENTE, USADO, ANULADO]
          example: PENDIENTE
        - name: id_conductor
          in: query
          description: |
            Filtra por conductor. Ignorado para un conductor puro (sin `jefe-area`,
            `administrador` ni `super-admin`), que siempre queda restringido a su propio
            `id_conductor` sin importar este parámetro.
          schema: { type: integer }
          example: 1
        - name: per_page
          in: query
          description: Cantidad de resultados por página.
          schema: { type: integer, default: 10 }
          example: 10
        - name: page
          in: query
          description: |
            Número de página a obtener (paginación estándar del listado, se resuelve
            automáticamente a partir de este parámetro). Si se omite, se asume `1`.
          schema: { type: integer, default: 1, minimum: 1 }
          example: 1
      responses:
        '200':
          description: Listado paginado de vales de combustible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValePaginada'
              example:
                current_page: 1
                data:
                  - id: 2
                    nro_vale: 2
                    gestion: 2026
                    fecha_emision: "2026-08-14 21:42"
                    fecha_vencimiento: "2026-08-14 22:42"
                    litros: "30.00"
                    precio: "6.97"
                    id_vehiculo: 1
                    id_conductor: 1
                    id_grifo: 2
                    estado_vale: PENDIENTE
                    id_tipo_combustible: 3
                    id_user: 2
                    nro: "000002/2026"
                    fecha_emision_f: "14/08/2026"
                    fecha_vencimiento_f: "14/08/2026"
                    vehiculo:
                      id: 1
                      nro_placa: 148-JLK
                      codigo: PM-TRA-0001
                      marca: Volvo
                      modelo: null
                      fotografia: vehiculos/CaPy03xlQuLTZhLkBPlZi0QlXmfDKkDpmwjgd5BS.jpg
                      url_fotografia: http://localhost:8000/storage/vehiculos/CaPy03xlQuLTZhLkBPlZi0QlXmfDKkDpmwjgd5BS.jpg
                    conductor:
                      id: 1
                      persona:
                        id: 1
                        nombres: Juan
                        paterno: García
                        materno: López
                        ci: "1234567"
                        foto: conductores/EmtgAtJKBQlpLmnxjY7W4VLw91eXG4Ct6e97IARI.jpg
                        fecha_nacimiento: "1990-05-15T04:00:00.000000Z"
                        foto_url: http://localhost:8000/storage/conductores/EmtgAtJKBQlpLmnxjY7W4VLw91eXG4Ct6e97IARI.jpg
                        edad: 36
                        f_nacimiento_formatted: "15/05/1990"
                        nombre_completo: Juan García López
                    grifo:
                      id: 2
                      razon_social: Gasolinera Central S.A.
                      ciudad: La Paz
                  - id: 1
                    nro_vale: 1
                    gestion: 2026
                    fecha_emision: "2026-08-14 21:19"
                    fecha_vencimiento: "2026-08-15 21:19"
                    litros: "25.00"
                    precio: "6.96"
                    id_vehiculo: 1
                    id_conductor: 1
                    id_grifo: 2
                    estado_vale: PENDIENTE
                    id_tipo_combustible: 3
                    id_user: 2
                    nro: "000001/2026"
                    fecha_emision_f: "14/08/2026"
                    fecha_vencimiento_f: "15/08/2026"
                    vehiculo:
                      id: 1
                      nro_placa: 148-JLK
                      codigo: PM-TRA-0001
                      marca: Volvo
                      modelo: null
                      fotografia: vehiculos/CaPy03xlQuLTZhLkBPlZi0QlXmfDKkDpmwjgd5BS.jpg
                      url_fotografia: http://localhost:8000/storage/vehiculos/CaPy03xlQuLTZhLkBPlZi0QlXmfDKkDpmwjgd5BS.jpg
                    conductor:
                      id: 1
                      persona:
                        id: 1
                        nombres: Juan
                        paterno: García
                        materno: López
                        ci: "1234567"
                        foto: conductores/EmtgAtJKBQlpLmnxjY7W4VLw91eXG4Ct6e97IARI.jpg
                        fecha_nacimiento: "1990-05-15T04:00:00.000000Z"
                        foto_url: http://localhost:8000/storage/conductores/EmtgAtJKBQlpLmnxjY7W4VLw91eXG4Ct6e97IARI.jpg
                        edad: 36
                        f_nacimiento_formatted: "15/05/1990"
                        nombre_completo: Juan García López
                    grifo:
                      id: 2
                      razon_social: Gasolinera Central S.A.
                      ciudad: La Paz
                first_page_url: "http://localhost:8000/api/v1/vales?page=1"
                from: 1
                last_page: 1
                last_page_url: "http://localhost:8000/api/v1/vales?page=1"
                links:
                  - url: null
                    label: "&laquo; Anterior"
                    page: null
                    active: false
                  - url: "http://localhost:8000/api/v1/vales?page=1"
                    label: "1"
                    page: 1
                    active: true
                  - url: null
                    label: "Siguiente &raquo;"
                    page: null
                    active: false
                next_page_url: null
                path: "http://localhost:8000/api/v1/vales"
                per_page: 10
                prev_page_url: null
                to: 2
                total: 2
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Vales]
      summary: Emitir un vale de combustible
      description: |
        Emite un vale de combustible (mismo comportamiento que el módulo web). Pensado para el
        **jefe de área** desde la app.

        **Autorización:** requiere el permiso `vales.crear` — lo tienen `jefe-area`,
        `administrador` y `super-admin`; un `conductor` recibe `403`. (En el sistema web este
        control es sólo de interfaz; en la API se aplica en el servidor.)

        Se calculan/asignan en el servidor y **no deben enviarse**: `nro_vale`, `gestion`,
        `fecha_emision`, `fecha_vencimiento` (según los parámetros de la empresa), `id_user`
        (usuario autenticado), `id_tipo_combustible` (se toma del vehículo) y `estado_vale`
        (siempre nace `PENDIENTE`).

        **Restricción del jefe de área:** sólo puede emitir vales para vehículos con una
        asignación de área activa/provisional a alguna de las áreas que tiene a cargo; en caso
        contrario responde `422` con el error en `id_vehiculo`. Administradores y super-admin
        no tienen esta restricción.

        El `id_conductor` normalmente es el conductor titular del vehículo — disponible en
        `GET /parametros/colecciones` → `vehiculos[].id_conductor` — pero se acepta cualquier
        conductor existente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValeStoreRequest'
            example:
              id_vehiculo: 8
              id_conductor: 45
              id_grifo: 2
              litros: 30
              precio: 6.97
      responses:
        '201':
          description: Vale emitido exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValeStoreResponse'
              example:
                message: "Vale #000003/2026 registrado exitosamente."
                data:
                  id: 3
                  nro_vale: 3
                  gestion: 2026
                  fecha_emision: "2026-09-03 10:15"
                  fecha_vencimiento: "2026-09-04 10:15"
                  litros: "30.00"
                  precio: "6.97"
                  id_vehiculo: 8
                  id_conductor: 45
                  id_grifo: 2
                  estado_vale: PENDIENTE
                  id_tipo_combustible: 1
                  id_user: 2
                  nro: "000003/2026"
                  fecha_emision_f: "03/09/2026 10:15"
                  fecha_vencimiento_f: "04/09/2026 10:15"
                  vehiculo: { id: 8, nro_placa: 1234-ABC, codigo: VOLQ-08 }
                  conductor: { id: 45, persona: { id: 45, nombres: Juan, paterno: Pérez, materno: Quispe } }
                  grifo: { id: 2, razon_social: Gasolinera Central S.A., ciudad: La Paz }
                  tipoCombustible: { id: 1, tipo_combustible: Diésel }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: El usuario autenticado no tiene el permiso `vales.crear`.
          content:
            application/json:
              example:
                message: No tiene permiso para emitir vales.
        '422':
          description: |
            Error de validación, o el jefe de área intenta emitir un vale para un vehículo
            fuera de sus áreas a cargo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              examples:
                validacion:
                  summary: Campos faltantes o inválidos
                  value:
                    success: false
                    message: Los datos enviados no son válidos.
                    errors:
                      litros: ["El campo litros es obligatorio."]
                      id_grifo: ["El campo estación de servicio es obligatorio."]
                vehiculo_fuera_de_area:
                  summary: Vehículo fuera del área del jefe
                  value:
                    success: false
                    message: Los datos enviados no son válidos.
                    errors:
                      id_vehiculo: ["Sólo puede emitir vales para vehículos asignados a su área."]
        '500':
          description: Error inesperado al emitir el vale.
          content:
            application/json:
              example:
                message: Error al registrar el vale.
                error: "Detalle técnico interno del error (varía según la causa)."
  /vales/{vale}:
    get:
      tags: [Vales]
      summary: Ver el detalle de un vale
      description: |
        Devuelve un vale con sus relaciones (`vehiculo`, `conductor.persona`, `grifo`,
        `tipoCombustible`, `user`).

        **Restricción de acceso:** un `conductor` sólo puede ver sus propios vales
        (`id_conductor` = persona del usuario); intentar ver el de otro responde `403`.
      parameters:
        - name: vale
          in: path
          required: true
          description: ID del vale.
          schema: { type: integer }
          example: 3
      responses:
        '200':
          description: Detalle del vale.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Vale' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Un conductor intenta ver el vale de otro conductor.
          content:
            application/json:
              example:
                message: No tienes permiso para ver este vale.
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido en `POST /auth/login`. Enviar como
        `Authorization: Bearer {access_token}`.
  responses:
    Unauthorized:
      description: No autenticado, token ausente/inválido/expirado.
      content:
        application/json:
          example:
            message: Unauthenticated.
    NotFound:
      description: El recurso solicitado no existe.
      content:
        application/json:
          example:
            message: "No query results for model [App\\Models\\OperacionDiaria] 999"
  schemas:
...
    ValidationErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: "Los datos enviados no son válidos." }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
          example:
            campo: ["mensaje de error"]

    # ── Autenticación ─────────────────────────────────────────────────
...
    ValeVehiculo:
      type: object
      description: Vehículo asociado al vale, tal como se devuelve en el listado (subconjunto de columnas).
      properties:
        id: { type: integer, example: 1 }
        nro_placa: { type: string, example: 148-JLK }
        codigo: { type: string, nullable: true, example: PM-TRA-0001 }
        marca: { type: string, nullable: true, example: Volvo }
        modelo: { type: string, nullable: true, example: null }
        fotografia: { type: string, nullable: true, example: vehiculos/CaPy03xlQuLTZhLkBPlZi0QlXmfDKkDpmwjgd5BS.jpg }
        url_fotografia: { type: string, nullable: true, example: http://localhost:8000/storage/vehiculos/CaPy03xlQuLTZhLkBPlZi0QlXmfDKkDpmwjgd5BS.jpg }
    ValeConductor:
      type: object
      description: Conductor asociado al vale, con su relación `persona` (subconjunto de columnas).
      properties:
        id: { type: integer, example: 1 }
        persona:
          type: object
          properties:
            id: { type: integer, example: 1 }
            nombres: { type: string, example: Juan }
            paterno: { type: string, example: García }
            materno: { type: string, nullable: true, example: López }
            ci: { type: string, nullable: true, example: "1234567" }
            foto: { type: string, nullable: true, example: conductores/EmtgAtJKBQlpLmnxjY7W4VLw91eXG4Ct6e97IARI.jpg }
            fecha_nacimiento: { type: string, format: date-time, nullable: true, example: "1990-05-15T04:00:00.000000Z" }
            foto_url: { type: string, nullable: true, example: http://localhost:8000/storage/conductores/EmtgAtJKBQlpLmnxjY7W4VLw91eXG4Ct6e97IARI.jpg }
            edad: { type: integer, nullable: true, description: Calculada a partir de `fecha_nacimiento`., example: 36 }
            f_nacimiento_formatted: { type: string, nullable: true, description: "`fecha_nacimiento` formateada como `d/m/Y`.", example: "15/05/1990" }
            nombre_completo: { type: string, description: "Concatenación de `nombres`, `paterno` y `materno`.", example: Juan García López }
    ValeGrifo:
      type: object
      description: Estación de servicio donde se emitió el vale (subconjunto de columnas).
      properties:
        id: { type: integer, example: 2 }
        razon_social: { type: string, example: Gasolinera Central S.A. }
        ciudad: { type: string, nullable: true, example: La Paz }
    Vale:
      type: object
      properties:
        id: { type: integer, example: 2 }
        nro_vale: { type: integer, description: Correlativo del vale dentro de la gestión., example: 2 }
        gestion: { type: integer, example: 2026 }
        fecha_emision: { type: string, example: "2026-08-14 21:42" }
        fecha_vencimiento: { type: string, nullable: true, example: "2026-08-14 22:42" }
        litros: { type: string, example: "30.00" }
        precio: { type: string, example: "6.97" }
        id_vehiculo: { type: integer, example: 1 }
        id_conductor: { type: integer, example: 1 }
        id_grifo: { type: integer, example: 2 }
        estado_vale:
          type: string
          enum: [PENDIENTE, USADO, ANULADO]
          example: PENDIENTE
        id_tipo_combustible: { type: integer, example: 3 }
        id_user: { type: integer, description: Usuario que registró el vale., example: 2 }
        nro: { type: string, description: "`nro_vale` formateado con ceros a la izquierda (6 dígitos) seguido de `/gestion`.", example: "000002/2026" }
        fecha_emision_f: { type: string, description: "`fecha_emision` formateada como `d/m/Y`.", example: "14/08/2026" }
        fecha_vencimiento_f: { type: string, nullable: true, description: "`fecha_vencimiento` formateada como `d/m/Y`.", example: "14/08/2026" }
        vehiculo: { $ref: '#/components/schemas/ValeVehiculo' }
        conductor: { $ref: '#/components/schemas/ValeConductor' }
        grifo: { $ref: '#/components/schemas/ValeGrifo' }
    ValeStoreRequest:
      type: object
      description: |
        Cuerpo esperado por `POST /vales`. `nro_vale`, `gestion`, `fecha_emision`,
        `fecha_vencimiento`, `id_user`, `id_tipo_combustible` y `estado_vale` se resuelven en
        el servidor y no deben enviarse.
      required: [id_vehiculo, id_conductor, id_grifo, litros, precio]
      properties:
        id_vehiculo:
          type: integer
          description: Debe existir en `vehiculo`. Para un `jefe-area`, debe estar asignado a una de sus áreas a cargo.
          example: 8
        id_conductor:
          type: integer
          description: Debe existir en `conductor`. Normalmente el conductor titular del vehículo (`vehiculos[].id_conductor` de `GET /parametros/colecciones`).
          example: 45
        id_grifo:
          type: integer
          description: Estación de servicio. Debe existir en `grifo` (ver `GET /parametros/colecciones` → `estaciones_servicio`).
          example: 2
        litros:
          type: number
          format: float
          minimum: 0.01
          maximum: 9999.99
          example: 30
        precio:
          type: number
          format: float
          minimum: 0.01
          maximum: 99999.99
          description: Precio por litro.
          example: 6.97
    ValeStoreResponse:
      type: object
      properties:
        message: { type: string, example: "Vale #000003/2026 registrado exitosamente." }
        data:
          allOf: [{ $ref: '#/components/schemas/Vale' }]
          description: El vale creado, con `vehiculo`, `conductor.persona`, `grifo` y `tipoCombustible` anidados.
    ValePaginada:
      type: object
      description: Estructura estándar de paginación, conservando los filtros de query actuales en los links de página.
      properties:
        current_page: { type: integer, example: 1 }
        data:
          type: array
          items: { $ref: '#/components/schemas/Vale' }
        first_page_url: { type: string, example: "http://localhost:8000/api/v1/vales?page=1" }
        from: { type: integer, nullable: true, example: 1 }
        last_page: { type: integer, example: 1 }
        last_page_url: { type: string, example: "http://localhost:8000/api/v1/vales?page=1" }
        links:
          type: array
          description: Lista de enlaces de paginación (anterior, cada número de página, siguiente), lista para renderizar controles de paginación.
          items:
            type: object
            properties:
              url: { type: string, nullable: true, example: "http://localhost:8000/api/v1/vales?page=1" }
              label: { type: string, example: "1" }
              page: { type: integer, nullable: true, example: 1 }
              active: { type: boolean, example: true }
        next_page_url: { type: string, nullable: true, example: null }
        path: { type: string, example: "http://localhost:8000/api/v1/vales" }
        per_page: { type: integer, example: 10 }
        prev_page_url: { type: string, nullable: true, example: null }
        to: { type: integer, nullable: true, example: 2 }
        total: { type: integer, example: 2 }
...